+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "android-attestation"]
-	path = android-attestation-root
+	path = warden-roboto
 	url = ../android-attestation.git
