+++ gradle.properties
@@ -1,3 +1,3 @@
 jdk.version=17
-artifactVersion=2.0.2
-androidAttestationVersion=1.5.2
+artifactVersion=2.1.0
+androidAttestationVersion=1.6.0
