+++ sample/backend/gradle.properties
@@ -1,5 +1,5 @@
 ktor_version=2.2.4
-kotlin_version=1.9.10
+kotlin_version=2.0.0
 logback_version=1.2.11
 kotlin.code.style=official
-attestation_version=2.0.0
+attestation_version=2.1.0
