+++ warden/src/test/java/JavaInteropTest.java
@@ -1,5 +1,5 @@
 import at.asitplus.attestation.AttestationService;
-import at.asitplus.attestation.DefaultAttestationService;
+import at.asitplus.attestation.Warden;
 import at.asitplus.attestation.IOSAttestationConfiguration;
 import at.asitplus.attestation.KeyAttestation;
 import at.asitplus.attestation.android.AndroidAttestationChecker;
@@ -26,7 +26,7 @@ public class JavaInteropTest {
 
     public static void testDefaults() {
         Assertions.assertThrows(AndroidAttestationException.class, () -> {
-                    new DefaultAttestationService(
+                    new Warden(
                             new AndroidAttestationConfiguration.Builder(new AndroidAttestationConfiguration.AppData(
                                     "at.asitplus.attestation-example", Collections.emptyList())).build(),
                             new IOSAttestationConfiguration(new IOSAttestationConfiguration.AppData(
@@ -36,7 +36,7 @@ public class JavaInteropTest {
                 "No signature digests specified");
 
         Assertions.assertThrows(AndroidAttestationException.class, () -> {
-                    new DefaultAttestationService(
+                    new Warden(
                             new AndroidAttestationConfiguration.Builder(new AndroidAttestationConfiguration.AppData("at.asitplus.attestation-example",
                                     new ArrayList<>()
                             )).build(),
@@ -48,7 +48,7 @@ public class JavaInteropTest {
                 "No signature digests specified");
 
         Assertions.assertThrows(AndroidAttestationException.class, () -> {
-                    new DefaultAttestationService(
+                    new Warden(
                             new AndroidAttestationConfiguration.Builder(new AndroidAttestationConfiguration.AppData("at.asitplus.attestation-example",
                                     new ArrayList<>(),
                                     10)
@@ -61,7 +61,7 @@ public class JavaInteropTest {
                 "No signature digests specified");
 
         Assertions.assertThrows(AndroidAttestationException.class, () -> {
-                    new DefaultAttestationService(
+                    new Warden(
                             new AndroidAttestationConfiguration.Builder(new AndroidAttestationConfiguration.AppData("at.asitplus.attestation-example",
                                     new ArrayList<>(),
                                     10,
@@ -76,7 +76,7 @@ public class JavaInteropTest {
                 "No signature digests specified");
 
         Assertions.assertThrows(AndroidAttestationException.class, () -> {
-                    new DefaultAttestationService(
+                    new Warden(
                             new AndroidAttestationConfiguration.Builder(new AndroidAttestationConfiguration.AppData("at.asitplus.attestation-example",
                                     new ArrayList<>()
                             )).build(),
@@ -91,7 +91,7 @@ public class JavaInteropTest {
     }
 
     public static void testAttestationCallsJavaFriendliness() throws NoSuchAlgorithmException {
-        AttestationService service = new DefaultAttestationService(
+        AttestationService service = new Warden(
                 new AndroidAttestationConfiguration.Builder(new AndroidAttestationConfiguration.AppData("at.asitplus.attestation-example",
                         Arrays.asList(new byte[][]{new byte[]{0, 1, 2, 3, 4, 5, 6, 7, 8}}))
                 ).build(),
